backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==4.2.0
textblob==0.17.1
